+++ changelog/RELEASE-NOTES.md
@@ -1,1 +1,2 @@
 - Made it easier to differentiate the search & selection background colors for all **61** themes. <sup><sup>Some days I question my current life choices....</sup></sup>
+- Added localization for all people hanging out in the Russian Federation. Thank you @Dragon-0609 for the translations!